Separate code review settings into a separate resource.
Since the code review settings reference teams and users but the teams
reference repositories we need to split the code review settings into a
separate resource that can be created/updated after both the repository
and team are created.

This ends up being a bit of a weird resource since the config syntax
allows more than one to be set for the same repo but in reality this won't
work since they'll all fight one another to update the same data.