Generic structure for integration resources
Beanstalk has a number of different integration types that each have their
own attributes. Since Terraform lacks a way to represent this generically
we'll instead have a number of separate Terraform resources in the UI that
map to the same underlying implementation.

Currently the schema functions are stubbed and do nothing.