fix: Force new resource on Integration credential change
While not strictly required by the Beanstalk API, forcing a new resource
when the credentials and endpoints change should lead to better failure
states.

On creation or update of the JIRA and Hipchat Integrations, Beanstalk
queries the target service to validate that:
- the credentials have the correct level of access
- the Hipchat rooms or JIRA projects exist

If either fails then the API operation fails with a 422 HTTP error.